+++ libs/dot-components/src/lib/ctm/index.ts
@@ -1,2 +1,3 @@
 export { DotProgressionBoard } from './progression-board/ProgressionBoard';
 export { DotProgressionBoardLegend } from './progression-board/ProgressionBoardLegend';
+export { DotProgressionBoardDrawer } from './progression-board/ProgressionBoardDrawer';