+++ libs/dot-components/src/lib/ctm/progression-board/ProgressionBoard.stories.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { Meta, Story } from '@storybook/react/types-6-0';
 import { DotProgressionBoard, ProgressionBoardProps } from './ProgressionBoard';
-import samplePhases from './sampleData';
+import samplePhases from './sample-data/sampleData';
 
 export default {
   title: 'Experimental/ProgressionBoard',
@@ -14,8 +14,4 @@ export default {
 
 export const Default: Story<ProgressionBoardProps> = (args) => (
   <DotProgressionBoard {...args} />
-);
-
-export const withDrawer: Story<ProgressionBoardProps> = (args) => (
-  <DotProgressionBoard displayDrawer={true} {...args} />
 );