+++ libs/dot-components/src/lib/ctm/workitem/WorkItemTooltip.styles.ts
@@ -1,6 +1,14 @@
 import styled, { css } from 'styled-components';
+import { Typography } from '@material-ui/core';
 
 export const rootClassName = 'wi-tooltip';
+
+export const StyledTypography = styled(Typography)`
+  ${({ theme }) => css`
+    color: ${theme.palette.layer.n0};
+    margin-bottom: 0;
+  `}
+`;
 
 export const StyledTooltipTitle = styled.div`
   ${({ theme }) => css`